refactor app.js (part 2) - extract implementatinoLoop
